requirements.txt
flask
flask-sqlalchemy
nltk
scikit-learn
numpy
pandas
spacy
python-dotenv
requests
PyJWT